Changed double to float.

+++ train-controller/src/main/java/hu/bme/mit/train/controller/TrainControllerImpl.java
@@ -7,7 +7,7 @@ public class TrainControllerImpl implements TrainController {
 	private int step = 0;
 	private int referenceSpeed = 0;
 	private int speedLimit = 0;
-	private double acceleration = 0;
+	private float acceleration = 0;
 
 	@Override
 	public void followSpeed() {
@@ -48,7 +48,7 @@ public class TrainControllerImpl implements TrainController {
 	}
 
 	@Override
-	public void setAcceleration(double acceleration) {
+	public void setAcceleration(float acceleration) {
 		this.acceleration = acceleration;
 	}
